+++ src/data/Numbers.kt
@@ -4,6 +4,7 @@ data class Numbers(
 	val numbers: List<Int> = emptyList(),
 	val weight: List<Double> = emptyList(),
 	val target: Int = 0,
+	val numBins: Int = 0,
 ) {
 	companion object {
 		fun new(name: String): Numbers? {
@@ -12,6 +13,7 @@ data class Numbers(
 				numbers = data["numbers"].toIntList(),
 				weight = data["weight"].toDoubleList(),
 				target = data["target"].parseInt(),
+				numBins = data["numBins"].parseInt(),
 			)
 		}
 		fun n(name: String): Int {